+++ R/RecreateLibrary.R
@@ -14,7 +14,7 @@
 #' @details A typical workflow is as follows:
 #' Run the \code{SavePackageNames()} command on an older version of \R
 #' to print to a text file a complete list of names for packages located under your current \R library tree(s).
-#' If desired, uninstall the older version of \R.
+#' Uninstall the older version of \R if it is no longer needed.
 #' On a freshly installed version of \R, with the \pkg{inlmisc} package available,
 #' run the \code{RecreateLibrary()} command to download and install the packages listed in the package-list text file.
 #'
@@ -49,7 +49,7 @@ SavePackageNames <- function(file="packagelist.txt", lib=NULL) {
             with(R.version, sprintf("# R version: %s.%s (%s-%s-%s)", major, minor, year, month, day)),
             sprintf("# Platform: %s", R.version$platform),
             sprintf("# User: %s", Sys.info()["user"]))
-  m <- rbind(matrix(meta), pkgs)
+  m <- rbind(matrix(meta), sort(unique(pkgs)))
   utils::write.table(m, file, quote=FALSE, row.names=FALSE, col.names=FALSE)
   cat(sprintf("Package list written to: \"%s\"\n", normalizePath(path.expand(file))))