+++ R/AddGradientLegend.R
@@ -11,6 +11,9 @@
 #' @param labels 'logical' or 'character'.
 #'   Can either be a logical value specifying whether (numerical) annotations are to be made at the tickmarks,
 #'   or a character or expression vector of labels to be placed at the tickpoints.
+#' @param scientific 'logical'.
+#'   Indicates if labels should be formatted for scientific notation,
+#'   see \code{\link{ToScientific}} for details.
 #' @param title 'character'.
 #'   Title to be placed at the top of the legend.
 #' @param loc 'character'.
@@ -32,16 +35,19 @@
 #' @export
 #'
 #' @examples
-#' plot(NA, xlim = c(0, 100), ylim = c(-10, 10), xlab = "x", ylab = "y")
+#' plot(NA, xlim = c(0, 100), ylim = c(-10, 10), xlab = "", ylab = "")
 #' breaks <- 0:200
 #' AddGradientLegend(breaks, rainbow, title = "Title")
 #' AddGradientLegend(breaks, rainbow, title = "Title", inset = c(0.2, 0.1))
+#' breaks <- seq(0, 2e+06, length.out = 5)
 #' AddGradientLegend(breaks, rainbow, loc = "topright", inset = 0.1)
-#' AddGradientLegend(breaks, rainbow, loc = "bottomright",
-#'                   inset = c(0.2, 0.1), strip.dim = c(1, 14))
+#' pal <- function(...) rev(rainbow(...))  # reverse colors in palette
+#' AddGradientLegend(breaks, pal, loc = "bottomright", inset = c(0.2, 0.1),
+#'                   scientific = TRUE, strip.dim = c(1, 14))
 #'
 
-AddGradientLegend <- function(breaks, pal, at=NULL, labels=TRUE, title=NULL,
+AddGradientLegend <- function(breaks, pal, at=NULL, labels=TRUE,
+                              scientific=FALSE, title=NULL,
                               loc=c("bottomleft", "topleft", "topright", "bottomright"),
                               inset=0, strip.dim=c(2, 8)) {
 
@@ -49,14 +55,19 @@ AddGradientLegend <- function(breaks, pal, at=NULL, labels=TRUE, title=NULL,
   on.exit(graphics::par(op))
 
   usr <- graphics::par("usr")
-  inches_in_pica <- 0.166667
+  inches_in_pica <- 0.16666667
   lwd <- 0.25
   cex <- 0.7
 
   if (is.null(at)) at <- pretty(breaks)
   if (is.character(labels) && length(labels) != length(at))
     stop("Arguments 'labels' and 'at' should be the same length.")
-  if (is.logical(labels) && labels) labels <- formatC(at, big.mark=",")
+  if (is.logical(labels) && labels) {
+    if (scientific)
+      labels <- ToScientific(at, lab.type="plotmath")
+    else
+      labels <- formatC(at, big.mark=",")
+  }
 
   if (length(strip.dim) == 1) strip.dim <- rep(strip.dim, 2)
   dx <- strip.dim[1] * inches_in_pica * (diff(usr[1:2]) / graphics::par("pin")[1])