+++ R/zzz.R
@@ -2,7 +2,7 @@
   if (interactive()) {
     ver <- read.dcf(file.path(lib, pkg, "DESCRIPTION"), "Version")
     packageStartupMessage(pkg, ": version: ", ver)
-    if (ver < 1 || grepl("\\.9000$", ver)) {
+    if (ver < 1 || grepl("\\.9[0-9]{3}$", ver)) {
       s <- paste("This information is preliminary or provisional and is subject to revision.",
                  "It is being provided to meet the need for timely best science.",
                  "The information has not received final approval by the U.S. Geological Survey (USGS) and",