+++ NEWS.md
@@ -1,5 +1,7 @@
 # inlmisc 0.2.7.9000
 
+- In `ToScientific` function, improve default value for digits argument.
+
 - Add `AddGradientLegend` function, used to add a continuous color gradient legend strip to a plot.
 
 - Remove `current_version` argument from `RecreateLibrary` function;